refactor: add the created service to CartService.ts
